src/03_movie_wiki/requirements.txt
kuzu==0.7.0
lancedb==0.16.0
python-dotenv==1.0.1
ell-ai[sqlite]==0.0.15
llama-index==0.12.2
llama-index-embeddings-openai~=0.3.0
llama-index-vector-stores-lancedb~=0.3.0
llama-index-graph-stores-kuzu~=0.6.0
cohere~=5.11.0